app/requirements.txt
Django
django-iris
irissqlcli
gunicorn
djangorestframework
markdown
django-filter